Fix Logic error
